+++ lib/nostr_mail.dart
@@ -9,6 +9,8 @@ export 'src/client.dart' show NostrMailClient;
 export 'src/constants.dart';
 export 'src/models/attachment_ref.dart';
 export 'src/models/email.dart' show Email;
+export 'src/models/email_summary.dart' show EmailSummary;
+export 'src/models/paginated_result.dart' show PaginatedResult;
 export 'src/models/encrypted_blob.dart';
 export 'src/models/gift_wrap_state.dart'
     show FailedGiftWrap, GiftWrapFailure, GiftWrapProgress, GiftWrapStage;